Give AM camera a rest when not in use.
When entering the attract screen, the AM camera is disabled.
When entering the 'script' screen, the AM camera is re-enabled.
This will hopefully mitigate any long-running problems that might occur
with using the color settings on the camera.

A minor update was also to not set the color format on the camera
every frame, but instead to only set it when it changes.

Release script was updated to exclude source files.

Admin screen (and oF code) was updated on both interactives
to allow the app to be re-started with a button.

Startup scripts were updated to work more seamlessly for different
setups (such as using nvm in development). This necessitates the
editing of Launchd PATH params as handled during bootstrapping.

Version bump to 1.0.1